nie testowany kod

+++ project-rest-api/src/main/java/com/project/controller/projekt/ProjectTeacherRestController.java
@@ -17,6 +17,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.net.URI;
+import java.util.Optional;
 import java.util.UUID;
 
 @RestController
@@ -69,7 +70,7 @@ public class ProjectTeacherRestController {
     }
 
     @DeleteMapping("/{projektId}")
-    public ResponseEntity<Void> deleteProjekt(@PathVariable Integer projektId,
+    public ResponseEntity<Void> deleteProjekt(@PathVariable Long projektId,
                                               @AuthenticationPrincipal User currentUser) {
         return projektService.getProjekt(projektId).map(p -> {
             if (!p.getTeacher().getTeacherId().equals(currentUser.getTeacher().getTeacherId())) {
@@ -81,7 +82,7 @@ public class ProjectTeacherRestController {
     }
 
     @GetMapping("/{projektId}/code")
-    public ResponseEntity<String> getJoinCode(@PathVariable Integer projektId,
+    public ResponseEntity<String> getJoinCode(@PathVariable Long projektId,
                                               @AuthenticationPrincipal User currentUser) {
         return projektService.getProjekt(projektId).map(p -> {
             if (!p.getTeacher().getTeacherId().equals(currentUser.getTeacher().getTeacherId())) {
